requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
django-rest-framework==0.1.0
Django==3.2.3
djangorestframework==3.12.4
httpie==2.4.0
idna==2.10
pip==21.1.1
psycopg2-binary==2.8.6
Pygments==2.9.0
PySocks==1.7.1
pytz==2021.1
requests-toolbelt==0.9.1
requests==2.25.1
setuptools==56.2.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
